MRZlib: route zlibDecompressStream through zlib-ng (zng_inflate*)
Complements the earlier zlibCompressStream-on-zlib-ng swap on this PR:
now both directions of MRMesh's public RFC 1950 / 1951 streaming API
run through zlib-ng, not just deflate. Motivation is per upstream
zlib-ng benchmarks and the AWS Open Source forks comparison:
decompression in zlib-ng is ~2-3x faster than stock zlib at the
inflate step (wider margin than the deflate speedup), driven by
SIMD-vectorised inflate_chunkcopy (SSE2/SSSE3/AVX2 on x86, NEON on
ARM). References in the upstream benchmark discussion
github.com/zlib-ng/zlib-ng/discussions/871 and the 2025 re-run at
github.com/zlib-ng/zlib-ng/issues/1486.

Concrete MeshLib consumers that inherit the speedup: decompressZip,
scene loaders that read .mrmesh blobs, and the round-trip verify
step in CompressManySmallFilesToZip (~15 MB decompressed per test
iteration, half its wallclock was inflate).

Implementation notes:

  - Port the existing Buffer<char> chunk-loop inflate code from
    MRZlib.cpp to MRZlibNg.cpp, replacing z_stream/inflate*/inflate
    with zng_stream/zng_inflate*/zng_inflate. Behaviour is identical:
    same 256 KiB chunks, same Z_NO_FLUSH/Z_STREAM_END handshake,
    same error codes (Z_* constants are shared between the two
    headers and have identical numeric values).

  - MRZlib.cpp had only the two zlibDecompressStream overloads left
    after the earlier compress move; with decompress now in
    MRZlibNg.cpp too, the old file has no live content. Delete it
    (and its ClCompile entries in MRMesh.vcxproj +
    MRMesh.vcxproj.filters). MRZlib.h stays -- it's the public API.

  - Round-trip remains compatible across forks. zlib-ng native's
    inflate reads any valid RFC 1950/1951 stream produced by stock
    zlib, libdeflate, zlib-ng, or any other DEFLATE encoder, so the
    parametrised ZlibDecompressTestFixture's pre-captured stock-zlib
    reference blobs continue to decompress correctly.

  - Stock zlib header <zlib.h> is no longer included by any MRMesh
    TU. MRMesh/CMakeLists.txt still does find_package(ZLIB) + link
    ZLIB::ZLIB because libzip transitively depends on it and the
    existing link line doesn't hurt; a cleanup PR can drop that
    link later if desired.
