+++ source/MRMesh/MRMesh.vcxproj
@@ -733,7 +733,6 @@
     <ClCompile Include="MRWatershedGraph.cpp" />
     <ClCompile Include="MRMeshOverhangs.cpp" />
     <ClCompile Include="MRZip.cpp" />
-    <ClCompile Include="MRZlib.cpp" />
     <ClCompile Include="MRZlibNg.cpp" />
     <ClCompile Include="MRObjectSave.cpp" />
     <ClCompile Include="MRRadiusMeasurementObject.cpp" />
